third_party/blink/web_tests/platform/generic/fast/table/cellIndex-of-cell-with-different-parents-expected.txt

Testcase for bug http://webkit.org/b/93738 . The testcase checks if the cellIndex of a table cell(td/th) is correct when the cell is inside different parents.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

The cellIndex of a table cell (td/th) must be -1 if its parent is other than a table row. Otherwise the cellIndex must be a value greater than or equal to 0.
PASS th_with_no_parent.cellIndex is -1
PASS td_with_no_parent.cellIndex is -1
PASS th_with_table_parent.cellIndex is -1
PASS td_with_table_parent.cellIndex is -1
PASS th_with_thead_parent.cellIndex is -1
PASS td_with_thead_parent.cellIndex is -1
PASS th_with_tbody_parent.cellIndex is -1
PASS td_with_tbody_parent.cellIndex is -1
PASS th_with_tfoot_parent.cellIndex is -1
PASS td_with_table_parent.cellIndex is -1
PASS first_cell_th_with_tr_parent.cellIndex is 0
PASS second_cell_td_with_tr_parent.cellIndex is 1
PASS successfullyParsed is true

TEST COMPLETE
